✨ feat(Config): Add option to configure maximum length of diff log
Added a new configuration option to allow users to set the maximum character length of the git diff log. The default value is 5000 characters. The option can be accessed and modified through the Config component in the user interface. The prompt message for generating a commit message has been updated to include information about the new configuration option and its default value.